Accepts words from user
You can type into the HTML form and “enter” the word in by hitting
space. The input part of this is working perfectly.

Now I’m starting to work on the word list generation part of the web
app, then I’ll test user input against the word that the website is
prompting with.