Generates a test, prompts a new testWord every time you submit a word
Kinda a funny problem to run into… Since I was getting my value as a
string and not a number, my ThreeDig function was working fine, until I
tried to do threeDig(x+1). (If x was 5, I would concatenate and then
call the function getting 051, instead of adding then call the function
getting 006). So that’s a lesson I learned today, and I won’t be
ignorant of the possibility of that mistake again.